Get rid of hostContext from CloudConfig, since it's always /solr, and update method signatures.
