+++ solr/core/src/test/org/apache/solr/cloud/ChaosMonkeyShardSplitTest.java
@@ -264,7 +264,7 @@ public class ChaosMonkeyShardSplitTest extends ShardSplitTest {
               "/admin/cores",
               reader,
               null,
-              new CloudConfig.CloudConfigBuilder("127.0.0.1", 8983, "solr").build());
+              new CloudConfig.CloudConfigBuilder("127.0.0.1", 8983).build());
       overseer.close();
       ElectionContext ec =
           new OverseerElectionContext(zkClient, overseer, address.replace("/", "_"));
